+++ Airship/AirshipAutomation/Source/InAppMessage/View/InAppMessageWebView.swift
@@ -31,6 +31,10 @@ struct InAppMessageWebView: View {
                     environment.onUserDismissed()
                 }
             )
+            .addBackground(
+                /// Add system background color by default - clear color will be parsed by the display content if it's set
+                color: displayContent.backgroundColor?.color ?? Color(.systemBackground)
+            )
             .zIndex(0)
 
             if self.isWebViewLoading {
@@ -94,12 +98,16 @@ struct WKWebViewRepresentable: UIViewRepresentable {
 
 
         private let parent: WKWebViewRepresentable
+        private let challengeResolver: ChallengeResolver
         let nativeBridge: NativeBridge
 
-        init(_ parent: WKWebViewRepresentable, actionRunner: NativeBridgeActionRunner) {
+        init(_ parent: WKWebViewRepresentable, actionRunner: NativeBridgeActionRunner, resolver: ChallengeResolver = .shared) {
             self.parent = parent
             self.nativeBridge = NativeBridge(actionRunner: actionRunner)
+            self.challengeResolver = resolver
+            
             super.init()
+            
             self.nativeBridge.nativeBridgeExtensionDelegate =
             self.parent.nativeBridgeExtension
             self.nativeBridge.forwardNavigationDelegate = self
@@ -131,6 +139,14 @@ struct WKWebViewRepresentable: UIViewRepresentable {
                 [weak webView] in
                 webView?.reload()
             }
+        }
+        
+        func webView(
+            _ webView: WKWebView,
+            respondTo challenge: URLAuthenticationChallenge)
+        async -> (URLSession.AuthChallengeDisposition, URLCredential?) {
+            
+            return await challengeResolver.resolve(challenge)
         }
 
         func performCommand(_ command: JavaScriptCommand, webView: WKWebView) -> Bool {