+++ AirshipExtensions/AirshipNotificationServiceExtension/Source/UANotificationServiceExtension.swift
@@ -166,10 +166,15 @@ open class UANotificationServiceExtension: UNNotificationServiceExtension {
     }
     
     private func download(url: URL) async throws -> (URL, URLResponse) {
+        let session = URLSession(
+            configuration: .default,
+            delegate: ChallengeResolver.shared,
+            delegateQueue: nil)
+        
         if #available(macOS 12.0, iOS 15.0, watchOS 8.0, *) {
-            return try await URLSession.shared.download(from: url)
+            return try await session.download(from: url)
         } else {
-            let (data, response) = try await URLSession.shared.data(from: url)
+            let (data, response) = try await session.data(from: url)
             let tmpUrl = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)
             try data.write(to: tmpUrl)
             return (tmpUrl, response)