+++ Thomas/Thomas/AppDelegate.swift
@@ -54,6 +54,13 @@ class AppDelegate: NSObject, UIApplicationDelegate {
             try await Airship.push.resetBadge()
         }
 
+        InAppAutomation.shared.inAppMessaging.themeManager.htmlThemeExtender = { message, theme in
+            if message.extras?.object?["squareview"]?.string == "true" {
+                theme.maxWidth = (min(UIScreen.main.bounds.size.width, UIScreen.main.bounds.size.height)-24*2)
+                theme.maxHeight = theme.maxWidth
+            }
+        }
+
         return true
     }